refactor: 중복 메시지 상수로 관리
