+++ src/main/java/com/programmers/board/controller/UserController.java
@@ -1,6 +1,6 @@
 package com.programmers.board.controller;
 
-import com.programmers.board.constant.SessionConst;
+import com.programmers.board.constant.AuthConst;
 import com.programmers.board.controller.response.PageResult;
 import com.programmers.board.controller.response.Result;
 import com.programmers.board.dto.UserDto;
@@ -52,7 +52,7 @@ public class UserController {
     @ResponseStatus(HttpStatus.NO_CONTENT)
     @PatchMapping("/users/{userId}")
     public void updateUser(@PathVariable("userId") Long userId,
-                           @SessionAttribute(name = SessionConst.LOGIN_USER_ID, required = false) Long loginUserId,
+                           @SessionAttribute(name = AuthConst.LOGIN_USER_ID, required = false) Long loginUserId,
                            @RequestBody @Valid UserUpdateRequest request) {
         checkLogin(loginUserId);
         userService.updateUser(
@@ -66,14 +66,14 @@ public class UserController {
     @ResponseStatus(HttpStatus.NO_CONTENT)
     @DeleteMapping("/users/{userId}")
     public void deleteUser(@PathVariable("userId") Long userId,
-                           @SessionAttribute(name = SessionConst.LOGIN_USER_ID, required = false) Long loginUserId) {
+                           @SessionAttribute(name = AuthConst.LOGIN_USER_ID, required = false) Long loginUserId) {
         checkLogin(loginUserId);
         userService.deleteUser(userId, loginUserId);
     }
 
     private void checkLogin(Long loginUserId) {
         if (loginUserId == null) {
-            throw new AuthenticationException(SessionConst.NO_LOGIN);
+            throw new AuthenticationException(AuthConst.NO_LOGIN);
         }
     }
 }
