+++ src/main/java/com/programmers/board/service/PostService.java
@@ -1,5 +1,6 @@
 package com.programmers.board.service;
 
+import com.programmers.board.constant.AuthConst;
 import com.programmers.board.domain.Post;
 import com.programmers.board.domain.User;
 import com.programmers.board.dto.PostDto;
@@ -15,6 +16,9 @@ import java.util.NoSuchElementException;
 
 @Service
 public class PostService {
+    private static final String NO_SUCH_USER = "존재하지 않는 회원입니다";
+    private static final String NO_SUCH_POST = "존재하지 않은 게시글입니다";
+    
     private final PostRepository postRepository;
     private final UserRepository userRepository;
 
@@ -41,7 +45,7 @@ public class PostService {
     @Transactional(readOnly = true)
     public PostDto findPost(Long postId) {
         Post post = postRepository.findByIdWithUser(postId)
-                .orElseThrow(() -> new NoSuchElementException("존재하지 않는 게시글입니다"));
+                .orElseThrow(() -> new NoSuchElementException(NO_SUCH_POST));
         return PostDto.from(post);
     }
 
@@ -63,17 +67,17 @@ public class PostService {
 
     private Post findPostOrElseThrow(Long postId) {
         return postRepository.findById(postId)
-                .orElseThrow(() -> new NoSuchElementException("존재하지 않는 게시글입니다"));
+                .orElseThrow(() -> new NoSuchElementException(NO_SUCH_POST));
     }
 
     private User findUserOrElseThrow(Long userId) {
         return userRepository.findById(userId)
-                .orElseThrow(() -> new NoSuchElementException("존재하지 않는 유저입니다"));
+                .orElseThrow(() -> new NoSuchElementException(NO_SUCH_USER));
     }
 
     private void checkAuthority(Post post, User loginUser) {
         if (!post.isWriter(loginUser)) {
-            throw new AuthorizationException("권한이 없습니다");
+            throw new AuthorizationException(AuthConst.NO_AUTHORIZATION);
         }
     }
 }
